Configuration: To enable Flash Attention, need to add `flash_attn`: true to model.json

Tested on windows 11 4900 cuda 12, can see an improvement on token speed
Without Flash Attention:
```
llama_new_context_with_model: flash_attn = 0
20240522 06:33:42.958000 UTC 7864 DEBUG [PrintTimings] PrintTimings: prompt eval time = 284.746ms / 80 tokens (3.559325 ms per token, 280.952146826 tokens per second) - llama_client_slot.cc:79
20240522 06:33:42.958000 UTC 7864 DEBUG [PrintTimings] PrintTimings:        eval time = 32301.621 ms / 297 runs   (108.759666667 ms per token, 9.19458500241 tokens per second)
 - llama_client_slot.cc:86
```

With Flash Attention:
```
llama_new_context_with_model: flash_attn = 1
20240522 06:31:41.109000 UTC 4872 DEBUG [PrintTimings] PrintTimings: prompt eval time = 632.817ms / 84 tokens (7.53353571429 ms per token, 132.739796813 tokens per second) - llama_client_slot.cc:79
20240522 06:31:41.109000 UTC 4872 DEBUG [PrintTimings] PrintTimings:        eval time = 47647.218 ms / 455 runs   (104.71916044 ms per token, 9.54935081414 tokens per second)
 - llama_client_slot.cc:86
```